requirements.txt
pycryptodome
psutil
